LayoutTests/fast/transforms/bounding-rect-zoom.html

<head>
    <title>Tests that getBoundingClientRect and getClientRects account for CSS zoom</title>
    <style>
    body {
        zoom: 3;
    }
    </style>
</head>

<body>
<div id="test" style="position:absolute; left:100px; top:100px; width:100px; height:100px;"></div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
const zoomFactor = 3;

test(() => {
    const element = document.getElementById("test");
    const bounds = element.getBoundingClientRect();

    assert_approx_equals(bounds.left, 300, 0.01);
    assert_approx_equals(bounds.top, 300, 0.01);
    assert_approx_equals(bounds.right, 600, 0.01);
    assert_approx_equals(bounds.bottom, 600, 0.01);
    assert_approx_equals(bounds.width, 300, 0.01);
    assert_approx_equals(bounds.height, 300, 0.01);
}, 'getBoundingClientRect() should account for CSS zoom');

test(() => {
    const element = document.getElementById("test");
    const rects = element.getClientRects();

    assert_approx_equals(rects[0].left, 300, 0.01);
    assert_approx_equals(rects[0].top, 300, 0.01);
    assert_approx_equals(rects[0].right, 600, 0.01);
    assert_approx_equals(rects[0].bottom, 600, 0.01);
    assert_approx_equals(rects[0].width, 300, 0.01);
    assert_approx_equals(rects[0].height, 300, 0.01);
}, 'getClientRects() should account for CSS zoom');
</script>
</body>
